\name{data}
\alias{data}
\docType{data}
\title{Example data}
\description{
This is a simulated dataset used to illustrate the performance of armDNA.
}
\usage{data(data)}
\format{
  A list containing three variables for 200 samples and 1000 DNA methylation marks.
  \describe{
    \item{\code{x}}{A vector for age of 200 samples}
    \item{\code{y}}{A matrix for DNA methylation profiles of 200 samples at 1000 marks}
    \item{\code{z}}{A matrix for 1 covariate of 200 samples}
  }
}

\keyword{dataset}
